pyproject.toml
[project]
name = "rag-fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
